test(vamm): Tidy tests
