+++ frame/vamm/Cargo.toml
@@ -28,6 +28,7 @@ scale-info = { version = "2.1", default-features = false, features = ["derive"] 
 serde = { version = '1', optional = true }
 
 [dev-dependencies]
+composable-tests-helpers = { path = "../composable-tests-helpers" }
 pallet-balances = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.22" }
 pallet-timestamp = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.22" }
 proptest = "1.0"
